requirements.txt
streamlit>=1.36.0
pandas>=2.2.0
gspread>=6.1.2
google-auth>=2.32.0
openpyxl>=3.1.0
